chore: rename sp.oidc.client to spcp.oidc.client
