+++ src/app/modules/form/public-form/__tests__/public-form.routes.spec.ts
@@ -13,7 +13,7 @@ import dbHandler from 'tests/unit/backend/helpers/jest-db'
 
 import { FormAuthType, FormStatus } from '../../../../../../shared/types'
 import * as AuthService from '../../../auth/auth.service'
-import { SpOidcClient } from '../../../spcp/sp.oidc.client'
+import { SpOidcClient } from '../../../spcp/spcp.oidc.client'
 import { PublicFormRouter } from '../public-form.routes'
 
 jest.mock('@opengovsg/myinfo-gov-client', () => ({
@@ -29,7 +29,7 @@ jest.mock('@opengovsg/myinfo-gov-client', () => ({
     .MyInfoAttribute,
 }))
 
-jest.mock('../../../spcp/sp.oidc.client')
+jest.mock('../../../spcp/spcp.oidc.client')
 
 jest.mock('@opengovsg/spcp-auth-client')
 const MockSpcpAuthClient = mocked(SPCPAuthClient, true)
