+++ src/app/modules/myinfo/__tests__/myinfo.routes.spec.ts
@@ -29,7 +29,7 @@ jest.mock('@opengovsg/spcp-auth-client')
 const MockAuthClient = mocked(SPCPAuthClient, true)
 
 // Avoid async refresh calls
-jest.mock('src/app/modules/spcp/sp.oidc.client.ts')
+jest.mock('src/app/modules/spcp/spcp.oidc.client.ts')
 
 jest.mock('@opengovsg/myinfo-gov-client', () => ({
   MyInfoGovClient: jest.fn().mockReturnValue({
