+++ src/app/modules/spcp/__tests__/spcp.controller.spec.ts
@@ -42,7 +42,7 @@ import {
   MOCK_TARGET,
 } from './spcp.test.constants'
 
-jest.mock('../sp.oidc.client')
+jest.mock('../spcp.oidc.client')
 
 jest.mock('../spcp.service')
 const MockSpcpService = mocked(SpcpService, true)
