+++ src/app/modules/spcp/__tests__/spcp.routes.spec.ts
@@ -31,7 +31,7 @@ import {
   MOCK_TARGET,
 } from './spcp.test.constants'
 
-jest.mock('../sp.oidc.client')
+jest.mock('../spcp.oidc.client')
 
 jest.mock('axios')
 const MockAxios = mocked(axios, true)
