+++ src/app/routes/api/v3/admin/forms/__tests__/admin-forms.feedback.routes.spec.ts
@@ -27,7 +27,7 @@ import { AdminFormsRouter } from '../admin-forms.routes'
 jest.mock('src/app/utils/limit-rate')
 
 // Avoid async refresh calls
-jest.mock('src/app/modules/spcp/sp.oidc.client.ts')
+jest.mock('src/app/modules/spcp/spcp.oidc.client.ts')
 
 const UserModel = getUserModel(mongoose)
 const EncryptFormModel = getEncryptedFormModel(mongoose)
