+++ src/app/routes/api/v3/admin/forms/__tests__/admin-forms.logic.routes.spec.ts
@@ -15,7 +15,7 @@ import { AdminFormsRouter } from '../admin-forms.routes'
 const UserModel = getUserModel(mongoose)
 
 // Avoid async refresh calls
-jest.mock('src/app/modules/spcp/sp.oidc.client.ts')
+jest.mock('src/app/modules/spcp/spcp.oidc.client.ts')
 
 const app = setupApp('/admin/forms', AdminFormsRouter, {
   setupWithAuth: true,
