+++ src/app/routes/api/v3/admin/forms/__tests__/admin-forms.twilio.routes.spec.ts
@@ -29,7 +29,7 @@ import { TwilioCredentials } from './../../../../../../services/sms/sms.types'
 jest.mock('src/app/utils/limit-rate')
 
 // Avoid async refresh calls
-jest.mock('src/app/modules/spcp/sp.oidc.client.ts')
+jest.mock('src/app/modules/spcp/spcp.oidc.client.ts')
 
 const MockAdminFormService = mocked(AdminFormService)
